implement contant table and reftable

+++ klang/include/value/klref.h
@@ -7,7 +7,7 @@
 #include <stdbool.h>
 
 typedef struct tagKlRefInfo {
-  unsigned int index;
+  uint8_t index;
   bool in_stack;
 } KlRefInfo;
